AI_Knowledge/STT_TTS_Evaluation/requirements.txt
azure-mgmt-resource
azure-mgmt-storage
azure-mgmt-web
azure-identity
azure-storage-blob
azure-data-tables
azure-functions
pandas